--- routes/userroutes.js
@@ -1,8 +1,0 @@
-// const express = require("express");
-// const router = express.Router();
-
-// const authController = require('../controllers/authcontroller');
-
-// router.post('/signup', authController.postSignup);
-
-// module.exports = router;
